requirements.txt
numpy
pandas
fastapi
uvicorn
python-multipart
torch-geometric
tqdm
scikit-learn
nibabel
SimpleITK
torch
torchvision
pyyaml
matplotlib
